+++ history.md
@@ -1,3 +1,9 @@
+[2026/03/27]
+- Updated
+  + 'null' 체크 우회를 '@SuppressWarnings'으로 변환 중.
+    + open.commons.core.utils.CsvUtils
+    
+
 [2026/03/25]
 - Updated
   + 'null' 체크 우회를 '@SuppressWarnings'으로 변환 중.